FIX readInt to work when sizeof(int) != 32

+++ tmc/ic/TMC5160/TMC5160.c
@@ -50,7 +50,7 @@ int32_t tmc5160_readInt(TMC5160TypeDef *tmc5160, uint8_t address)
 	data[0] = address;
 	tmc5160_readWriteArray(tmc5160->config->channel, &data[0], 5);
 
-	return ((uint32_t)data[1] << 24) | ((uint32_t)data[2] << 16) | (data[3] << 8) | data[4];
+	return ((uint32_t)data[1] << 24) | ((uint32_t)data[2] << 16) | ((uint32_t)data[3] << 8) | (uint32_t)data[4];
 }
 
 // Initialize a TMC5160 IC.
